requirements.txt
langchain
langchain-community
pypdf
sentence-transformers
faiss-cpu #vector database
ollama
streamlit
tqdm
numpy